rename wireup
