+++ lib/extensions.js
