+++ lib/index.js
@@ -42,6 +42,6 @@ if (typeof Realm != 'undefined') {
     throw new Error('Missing Realm constructor - please ensure RealmReact framework is included!');
 }
 
-require('./wireup')(realmConstructor);
+require('./extensions')(realmConstructor);
 
 module.exports = realmConstructor;
